requirements.txt
pytest==2.9.1
numpy==1.9.2
scipy==0.14.1
python-igraph==0.7.1.post6
pytest==2.9.1